Add rented rooms to the Aerie Guest House
This commit adds rented rooms to the island inn using the same system as
Marion and Barloque rooms. Additional textures are now available to
renters that match island decor. They are the default for island rooms,
but are also now available in the other rented rooms.

It also fixes a rare issue with textures - if the starting bed location
was randomly chosen to be the same number as the default, 1, the bed
would not be lifted from the floor or textured. Now that no longer
happens.